Feature/new llm by disease (#26)
* Update forecast models for changes to LLM

* Update to capture new LLM output by disease. fix an issue wth  stats standards breaking

* mpox fixes

* Revised LLM output added

* Update LLM v2 output to just dump a single file now that disease summaries are separate
(Claude) modify Deployhook so that it can be used by multiple methods.
fix env speeling error

* Incorrect data in airtable.
LLM Summaries to airtable from the llm summaries. Use summary-short as the updates
Netlify triggers to python 3.13

+++ workflows/public/public/resources/airtable.py
@@ -272,7 +272,8 @@ class AirtableResource( ResourceWithAirtableConfiguration):
                 df_clean[col] = df_clean[col].fillna(0)
                 # Convert numpy types to native Python types
                 if pd.api.types.is_integer_dtype(df_clean[col]):
-                    df_clean[col] = df_clean[col].astype('Int64').fillna(0)
+                    #df_clean[col] = df_clean[col].astype('Int64').fillna(0)
+                    df_clean[col] = df_clean[col].astype(float).fillna(0)
                 else:
                     df_clean[col] = df_clean[col].astype(float).fillna(0.0)
